backend/api-gateway/requirements.txt
fastapi==0.116.1
uvicorn[standard]
httpx==0.28.1
requests==2.32.4
PyJWT==2.10.1
python-decouple==3.8
